Split and extend tests.
