+++ test/rabbit_test.go
@@ -1,0 +1,110 @@
+package test
+
+import (
+	"context"
+	"sync"
+	"time"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	. "github.com/getkalido/rabbit-events"
+)
+
+var _ = Describe("Rabbit", func() {
+	Describe("RabbitBatcher", func() {
+		It("Should send messages after quiescence", func() {
+			messagesSent := make([]string, 0)
+			var l sync.Mutex
+			mockSender := func(ctx context.Context, m []string, x string, ri RabbitConfig, _ map[string]interface{}) {
+				l.Lock()
+				defer l.Unlock()
+				messagesSent = append(messagesSent, m...)
+			}
+			rb := &RabbitBatcher{QuiescenceTime: time.Nanosecond, MaxDelay: time.Hour, BatchSender: mockSender, Config: &RabbitIni{}}
+			go func() {
+				defer GinkgoRecover()
+				rb.Process()
+			}()
+			rb.QueueMessage("A")
+
+			time.Sleep(time.Millisecond * 40)
+
+			l.Lock()
+			defer l.Unlock()
+			Expect(messagesSent).To(HaveLen(1))
+
+			rb.Stop()
+		})
+
+		It("Should send messages before max timeout", func() {
+			callCount := 0
+			var l sync.Mutex
+
+			mockSender := func(ctx context.Context, m []string, x string, ri RabbitConfig, _ map[string]interface{}) {
+				l.Lock()
+				defer l.Unlock()
+				callCount++
+			}
+			rb := &RabbitBatcher{
+				QuiescenceTime: time.Millisecond * 4,
+				MaxDelay:       time.Millisecond * 5,
+				BatchSender:    mockSender,
+				Config:         &RabbitIni{},
+			}
+			go func() {
+				defer GinkgoRecover()
+				rb.Process()
+			}()
+
+			for k := 0; k < 15; k++ {
+				rb.QueueMessage("A")
+				time.Sleep(time.Millisecond * 1)
+			}
+
+			rb.Stop()
+
+			l.Lock()
+			defer l.Unlock()
+			Expect(callCount > 1).To(Equal(true))
+		})
+
+		It("Should send all messages ", func() {
+			messagesSent := make([]string, 0)
+			var l sync.Mutex
+
+			mockSender := func(ctx context.Context, m []string, x string, ri RabbitConfig, _ map[string]interface{}) {
+				l.Lock()
+				defer l.Unlock()
+				messagesSent = append(messagesSent, m...)
+			}
+			rb := &RabbitBatcher{
+				QuiescenceTime: time.Millisecond * 4,
+				MaxDelay:       time.Millisecond * 5,
+				BatchSender:    mockSender,
+			}
+			go func() {
+				defer GinkgoRecover()
+				rb.Process()
+			}()
+
+			var wg sync.WaitGroup
+			for k := 0; k < 15; k++ {
+				wg.Add(1)
+				go func() {
+					defer wg.Done()
+					rb.QueueMessage("A")
+				}()
+			}
+			wg.Wait()
+
+			time.Sleep(6 * time.Millisecond)
+
+			rb.Stop()
+
+			l.Lock()
+			defer l.Unlock()
+			Expect(messagesSent).To(HaveLen(15))
+		})
+	})
+})
